dcache-view: move username+password logic in view-file and file-metadata element
Motivation:

The attached function in Polymer is called whenever an element is attached
to the document. The view-file and file-metadata element (these are custom
dcache-view elements) makes use of this attached function to setup username
and password.

Unfornately, some browsers seem not to call the attached function when
attaching the view-file and file-metadata element to the DOM. This can be
traced down to the fact that those browsers seem not to catch-up with
mutiple creation and deletion of element (especially view-file) dynamically.
The side-effect of this is that user is never authenticated by dcache-view.

Modification:

Removed attached function and used the computed function to set the value
of the upw property for view-file and file-metadata element.

Result:

All browsers are happy

Target: trunk
Target: 1.0
Require-note: no
Require-book: no
Acked-by: Paul Millar<paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9550/