dcache-view: introduce hover context menu
Motivation:

To open or download a file in dcache-view, user will need to double click the file.
This feature was desirable because it enable a pathway to multiple selection of
files, that is, a single click on a file will select the file and a double click
will open or download it.

However, some user might find double-clicking for navigating a file "annoying".
Here is a solution that improve user-friendliness of dcache-view by providing a
contextual, which will only show up when a user mouse move over a file. And
the buttons in the contextual will response to a single click. For now, this hover
context menu contain two action buttons: open/download a file and view file
metadata button.

Modification:

1. Created a new element called hover-contextual element. As noted above, this
element has two buttons and more can be included later as new features are added.
2. Introduced two functions that will handle both mouse over and leave event in the
list-row element. Basically, these functions are called when the respective events
are triggered. During mouseover event, the displayed file size will be hidden and
the hover context menu will be shown in its place. And when mouse leave is triggered,
the file size is visible and the hover context menu will be removed.

Result:

Enhance the usability and friendliness of dcache-view

Target: trunk
Request: 1.0
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Worked-with: 2.16
Fixes: https://github.com/dCache/dcache/issues/2465

Reviewed at https://rb.dcache.org/r/9561/