dcache-view: add a check to the function responsible for downloading
Motivation:

In the config.js file used by dcache-view for configuration purpose.
One of the options available for configuration is the webdav endpoint
(that is, the full address of the webdav door). This webdav endpoint
is use by dcache-view to download file(s). The default configuration
assumed that both webdav and frontend services have the same hostname
and use the same protocol but with different port numbers (webdav: 2880
and frontend service: 3880). If any of these variable change from the
ones stated above, as mentioned earlier, this must be reflected in the
config.js. file.

As instructed in the dcache frontend.properties file for setting webdav
endpoint, "One could set this to something like https://example.org:443/"
and with strong emphases that this "MUST end with a slash if non-empty."

Since the file path always start with forward slash, concatenating both
the webdav endpoint and file path without checking can lead to error in
downloading.

Modification:

Add a check to the function that is responsible for downloading a file.
This check remove a forward slash if the webdav endpoint ends with one
because the file path always start with forward slash.

Result:

No visible change to the user.

Target: trunk
Request: 1.0
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Worked-with: 2.16

Reviewed at https://rb.dcache.org/r/9562/