dcache-view: add files upload functionality and feature
Motivation:

The upload feature is neccessary for any storage system like
dcache. The new user interface of dcache, called dcache-view
now implement the available facility inside dcache webdav for
file/s upload.

Modification:

1. Creates an equivalent of a class in javascript to handle the
upload.
2. The upload-file-button element is added to the selected-title
element as part of the tools that will be available for users. On
click a file dialog-box will pop-up, which will allow the user to
select file or multiple of files to be uploaded. When enter button
on the keyboard or open button on the file dialog-box is hit, an
object of upload class will be created for each file/s.
3. An upload-toast-file element will display the status of the file.

Result:

User can upload file/s into a directory. The uploaded file/s will
be added to the top of the list of files in that directory without
reloading the page.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9808/