dcache-view: re-define the process of directory listing
Motivation:

dCache-view list files in a directory by using page routing, which in turn
trigger an ajax request. When the response of this ajax request arrived,
polymer is not neccessarily notified about these changes. This make
rendering of the content of a directory misbehave and unpredictable.

Also, we need to adjust the listing element in other to show any live update
on the data without refreshing the whole page. This patch pave way for that
to be possible.

Modification:

1. create a new view-file element each time a request to list directory is made
2. remove dependency on page routing for listing files
3. pagination button, selected-title element and the index.html are adjusted
to accomdate the new structure.

Result:

Better and stable performance.

Target: trunk
Request: 1.0
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9767/