dcache-view: hide table header
Motivation:

The table header is an enhancement feature that show what is display on each
row of a listed directory. At some instant, a user might be required to login
in other to view the content of a directory OR the requested directory might
be empty. In either case, the table header should be hidden to the user.

Modification:

1. Change the parent id of the list-row-header element for easy accessibilty
and the corresponding css ref. is adjusted.
2. Add hidden and visible class to the theme css.
3. Call the list-row-header element inside view-file element to either hide or
make visible when appropriate.

Result:

Look less ugly to the user.

Target: trunk
Request: 1.0
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9810/