dcache-view: add create new directory feature
Motivation:

Implement the dcache restful-api for creating a new directory in dcache-view.

Modification:

A button is added to the selected-title element to handle creation of
new directory. When clicked a dialog box will be display. dCache-view have
a centralised dialog box, meaning that a dialog box can be call at the any-
time with different content. For creating a directory, the create-directory
element (see below for the repo and doc/demo page of this element) was used
as the content of the central dialog box.

The create dialog box content has an input textbox and two buttons. It
also respond to both escape and enter key. By default the input text
will be focused. The create event of this element will be fired by
clicking the create bottom or the enter keys. When this event is fired,
it will make an ajax call to dcache  by using dcache-namespace element.

<create-directory/>
    repo: https://github.com/dcache-elements/create-directory
    doc and demo: https://dcache-elements.github.io/create-directory/components/create-directory/

<dcache-namespace/>
    repo: https://github.com/dcache-elements/dcache-namespace
    doc and demo: https://dcache-elements.github.io/dcache-namespace/components/dcache-namespace/

Result:

End user can earsily create a new folder to a directory.

PS: please take a look at the above repos.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9874/

Reviewed at https://rb.dcache.org/r/9875/