dcache-view: add delete button for a file
Motivation:

One of the basic functions of data storage system is the abilty
to delete a file. dCache provide this facilty and current dCache
RESTful API recently include this in it namespace operation.

This patch use the implemented REST API call to provide the end user,
that is, dCache-view user the abilty to delete a file provided the
user have the right to do so.

Modification:

Add a delete button that will only be visible to files with 'REGULAR'
fileType. This button is added to the list of actions in the hover menu.
The delete button make use of dcache-namespace element to make the call
to dcache.

Inside dcache-namespace element, a delete method was created to make use
of http DELETE method.

Result:

User can now delete a file.

PS: please take a look at changes made to dcache-namespace element here:
<https://github.com/dcache-elements/dcache-namespace/commit/66db90a77aaf3c9ede61dbdd7569abb6185841b7>

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Albert Rossi <arossi@fnal.gov>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>

Reviewed at https://rb.dcache.org/r/9990/