dcache-view: ensure proper display of empty and error page
Motivation:

When a new file is uploaded or a new folder is created into an empty directory,
this will not be display to the end user immediately. An end User will need to
reload the page to see the new file or any new modification.

Also, when a directory contain one file and this file was deleted; an empty
directory is not display to the user to reflect the new changes. This applies
to when an error occured during the listing of a directory as well.

Modification & Result:

1. Add the appropriate class to the list-table-header element at the attached
and detached function of:
    a. directory-ls-error;
    b. empty-directory; and
    c. user-login-page
2. Make create, upload and delete features to be more response and display
proper new information on the fly without refreshing the page.

Target: trunk
Request: 1.1
Requires-notes: no
Requires-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/10038/

(cherry picked from commit 03d6a3375323320b9aa39469d3387878fa35783c)