dcache-view: add move file operation
Motivation

Moving file from one directory to another is part of standard
namespace operation. This patch added this feature to dcache-view.

Modification:

A move button was added. This will only appear when a user click
on a file. And If the button is clicked, a dialog box will be
displayed. The content of this dialog box is the newly created
element called move-file. The move-file element displays only
directories in the current view. Hence, any other file type that
is not a DIR will not be listed. Also, the selected file will not
be listed even if the file is a directory.

The displayed dialog box during move operation include a create
new directory button.

Update the third party dependecies the newiest version to avoid
conflict.

(See: https://github.com/dcache-elements/move-file for the move-file element.)

Result:

1. An interactive and easy to use move dialog box.
2. User can create a new directory on the fly without dismissing the
dialog.

Target: trunk
Requires-notes: no
Requires-book: no
Acked-by: Paul Millar <paul.millar@desy.de>