dcache-view: add rename file feature
Motivation:

Renaming of a file is part of basic function in a storage system.
This operation is already available in dcache restful service.

This patch expose the already existing feature of renaming a file
in dcache to dcache-view.

Modification:

1. add a rename button to the name column of the list-row element.
2. add a dialog box for renaming in the view-file element
3. this dialog respond to esc and enter key. The escape key to
dismiss the box and the enter to make a request for changing the
file name. The rename is done by using dcache-namespace element.

Result:

User can now rename a file.

Target: trunk
Requires-notes: no
Requires-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/10036/