dcache-view: hide main menu button base on page width
Motivation:

The main menu button always appear on top of dcache-view.
This menu button is meant to be used to display the left-hand
drawer when the page is 'narrow', that is, when the page width
becomes smaller (e.g on mobile screen).

Therefore, it is redundant to display this menu button when the
page width is wide enough.

Modification:

Added event listener (`paper-responsive-change`) that will be fired
based on the page width. When the page is narrow, the menu button
will be display and vice versa.

Result:

Provide a task focus user interface.

Target: trunk
Request: 1.2
Request: 1.1
Request: 1.0
Requires-notes: no
Requires-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/10053/