dcache-view: factor out directory rendering
Motivation

Rendering of the children of a directory is a major part of dcache-view.
Since this is mostly done based on user interactions and actions; these
acrions are responded to by different functions in different elements.
This have introduce some codes duplication that can be factored out.

Modification:

Create a method called ls. This method will be called to render the content
of a directory. If the path to the directory is supplied; the children of
the directory will be rendered, else the root directory will be displayed.

These following elements make use of this method:
    a. list-row, b. pagination-button and c. hover-contextual.

Result:

Reduce code duplication.

Target: trunk
Requires-notes: no
Requires-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/10108/