dcache-view: stay in current path after log in
Motivation:

When a user is required to login or the user itself clicked on
the login button, dcache-view will reload the page after the user
have been authenticated. Since dcache-view is using client side
routing system, the reload of the page will redirect to the root
directory; which might be undesirable to the user.

Modification:

Instead of reloading the page after successful login, use the new
rendering directory method to display the current path and notify
the user-loginout-button and the subproperty of CONFIG, which is
called isSomebody of the current status.

Result:

User can now stay on the current path after login.

NB: this patch rely on https://rb.dcache.org/r/10108/

Target: trunk
Requires-notes: no
Requires-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/10109/