dcache-view: build a customised context menu
Motivation:

As dcache-view is becoming more stabled fully fledged web app,
this require that the user interaction within the application
emulate the native behaviour. Meaning that interaction like a
right-click mouse operation should provide a limited set of
choices that are available in the current state, or context,
of the application.

At the present, dcache-view provide these choices in various
places within the application (like on mouse over of a file)
and this make design not so friendly nor familiar to end user.

Modification:

1. Prevent default right click context menu from displace.
2. Remove the context menu that is displayed on mouse over.
3. Create a namespace context menu that will be shown only
at namespace space, that is, at the directory listing page.
The content of the context menu or the actionable operation
will depend on where and what was right-clicked on.
4. Add a dialog box to the index page that will be use for
displaying the application context menu.
5. Response to right-click mouse operation on individual or
group of files and on the current directory.

Result:

Cleaner user interface and closer emulation of native behaviour.

Target: master
Requires-notes: no
Requires-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>

Reviewed at https://rb.dcache.org/r/10144/