dcache-view: upload files and directories using HTML5 drag&drop API
Motivation:

File(s) upload using HTML drag and drop API is a desirable
feature in a web app like dcache-view. This will enable user
to upload file by simply dragging the file (chrome & firefox
support drag and drop for folder) into the current directory
or a listed directory in the current view.

Modification:

1. Prevent the generic drag and drop on the page.

2. Allow and listen to the drop, dragenter, dragleave, dragend,
dragexit events in the namespace section of the page.

3. Attached the neccessary drag & drop events to the list-row
and pagination-button element. This will enable file(s) to be
dropped on a listed dir. Also, if the user dragged and stayed
on a listed directory for more than 2 seconds; this directory
will be open and the user drag operation will still continue.

4. Create two new elements namely `drag-enter-toast` & `dnd-upload`.
The drag-enter-toast element notify the user when it enter
into a drop zone. While dnd-upload element upload the files
when dropped.

5. Add a cental function that can be use to check the broswer type
and version.

Result:

User can now drag and drop their files (and even folders if the
browser support it) into dcache-view for uploading.

**NB: drag and drop of folder is not supported in Safari yet.**

Target: master
Requires-notes: no
Requires-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>

Reviewed at https://rb.dcache.org/r/10207/