dcache-view: move file(s) using HTML5 drag and drop API
Motivation:

Recently, drag and drop for uploading files was implemented.
This patch added one of the possible features of the drag and
drop API; which is the ability to drag and drop an element
within the web page. In namespace operation this will equate
to moving of files from one directory to another. Hence, this
patch implement this feature.

Modification:

Adds dragstart event to list-row element; this enable user to
be able to select and drag a single or multiple files if the
multiple selection toggle button is enable.

Readjust the drop event in the list-row element and the one
for the current view in other to accomodate the move operation.

Result:

User can now move file(s) from one directory to another by
simply selecting and dragging it (or them) into the destination
directory.

Target: master
Requires-notes: no
Requires-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/10218/