dcache-view: adjust to work with dcache 3.0
Major dcache-view version should be normalise to a
dcache version, that is,
    dcache-view 1.0.X -> dcache 2.16.x
    dcache-view 1.1.X -> dcache 3.0.x
    dcache-view 1.2.x -> dcache 3.1.x
Unfornately, some patches are in 1.1.x that contains features
that are not in dcache 3.0. For example: file deletion and
file mime type.

The patch remove file deletion and adjust file mime type in
other to work with dcache 3.0

Request: 1.1
Requires-notes: no
Requires-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/10262/