dcache-view: fire event when the error message change
Since errorMessage is a property of the user-authentication
element and whenever the errorMessage change, an event is fired.
This event is listen to by other element in other to properly
handle any error that occur during authentication.

To make the fired event more generic in this element, an observer
is added; which is observe changes in the errorMessage property.
The function attached to this observer will be called whenever
errorMessage changes. This in turn will fire an even which can be
listen to.

Therefore, no need to fire an event 'manaully' each time an error
message is set.

Target: master
Require-book: no
Require-notes: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/10302/