dcache-view: add email to list of user info and support gravatar
Motivation:

Email address (if available) is part of discoverable user information
in frontend service. If the user have a gravatar profile linked with
this email address; it will be nice to discover the gravatar profile
image of the user. dCache-view can use this image by displaying it
in-place of the user identicon (which is generated using the username)
on the user-profile page. (For more information on gravatar, see
http://en.gravatar.com/)

Modification:

1. Rename and readjust user-identicon element. The element main
function now is to generate a image data and the rendering part is
now moved to user-image element. The name changing of the element
from user-identicon to identicon-request is to reflect the current
function of the element.

2. Since implementaion of CORS by browser prevent direct acess to
gravatar profile. It become necessary to add iron-jsonp-library
to the list of iron-elements and this is mainly use in the newly
created element to make a request to the gravatar website.

3. Add email address is now added to the list of user information
stored in the sessionStorage object.

4. Create a new element called gravatar-request. This will be used
whenever the user email address or their equivalent md5 hash is
provided. A request will be make to check if there is an associated
gravatar profile with the email.

5. The user-image element will be responsible in making sure that
a profile image is displayed. This element check if gravatar-request
is sucess and generating an image link base on the email address and
if the request fail (or there is no email address provided), the
fallback option is use, which is the identicon.

Result:

User email address is added to the list of user information and a
gravatar image will be use if available.

Target: trunk
Request: 1.3
Requires-notes: no
Requires-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/10368/

(cherry picked from commit b94c529f22c848636750977a782993a637b794e6)