dcache-view: prevent file removal when dnd is cancelled
Motivation:

When a user drags a file or files and release the
mouse before droping it on any target, this result in
the file/s to be removed from the current display list
of files. The effect of this behaviour should should
equal to cancelling the move operation. Hence, the
current implementation is wrong.

Modification:

Add a check that will prevent the moving of files if the
destination path is the same as the source path.

Result:

Prevent file/s from been removed from the list and no more
display of unnecessary toast message.

Target: master
Request: 1.3
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/10657/