dcache-view: adjust toast message for move dnd
Motivation:

When a move op is initiated by drag and drop; if error
occurs like in the case of unauthorise user; the message
that appear in the toast is not clear.

Also, when the move is successful, the paths (both the
destination and source) are not clearly indicated in the
message.

Modification:

1. Update dcache-namespace version and resolve to the
lastest version. The new update provide useful error
message.

2. Create a div and append it to the paper-toast. This
div contain the source and destination path of the move
in a well styled format. And when the paper-toast is
closed, the div will be removed. By removing the div;
this ensured that the paper-toast can still be used by
other features.

3. The span tag is replaced with paper-icon-button with
close icon. This help to distinguish the message from
the button.

Result:

Better messsage for move operation for drag and drop.

Target: master
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>

Reviewed at https://rb.dcache.org/r/10658/