dcache-view: allow move operation in pagination-button
Motivation:

At the moment, only drag and drop for file upload is
allowed by pagination-button element. This feature
can be expected for moving files; like draging a file
from the view-file and drop it on one the directory
(depicted by the pagination-button) in the working
directory.

Modification:

1. Enable event retargeting for all the event listeners.
2. When a file is dropped on the pagination-button, a
checked is performed in the drop event to see if the
operation is a move or upload.

Result:

User can now drop a drag file onto the pagination button
to initiate a move operation.

Target: master
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>

Reviewed at https://rb.dcache.org/r/10656/