dcache-view: replace autoselect column on pools of group dialog cell info table
Motivation:

Vaadin-grid provides an option on the grid selection column
for it to add automatically the items of checked rows to
the list of selected items.

Use of this feature, however, does not play well with click/tap
events intercepted on a given row in order to take action
(in this case, open a pool dialog).  In particular, the
click or tap also checks the checkbox, a component which
is in the element's shadow dom and does not have an id.

Modification:

Borrow/adapt from the provider decorator the code which manages
selection columns on tables for which we also construct
a provider (for filtering and sorting on remote data
which is lazily loaded through infinite scrolling).

The provider decorator cannot be reused directly here
because that would also entail substitution for the direct
fetch of the table data the lazy-load mechanism, which in the
case of this component is overkill.   Hence the
provider decorator methods are adapted as private methods
of this component.

Result:

Clicking on the pool name to open the dialog no longer
checks the checkbox in the selection column.

Target: master
Acked-by: Olufemi
Require-notes: no
Require-book: no