dcache-view: fix column count computation on pool plots view
Motivation:

The distribution of columns in the grid is based on the square root
of the plot count up to a maximum of 8.  There is one exception:
when there is a single plot, it should be centered on the page.

There is a bug which currently centers up to 3 plots ( < 4), because
it is taking the Maximum of 1 and the square root of that number.

Modification:

Rewrite the conditions to exclude a single plot count from the
square root computation, and compare the square root to a minimum
of 2.

Result:

No overlapping in the center.

Target: master
Require-notes: no
Require-book: no
Acked-by: Olufemi