dcache-view (admin): correct parsing of pool name url
Motivation:

When displaying pool usage information, the request URL
is parsed in the reply in order to find the pool name.

This code, however, was not adjusted when changes were
made to the RESTful api.  The poolname is no longer
the last element in the URL path.

This is only apparent when the response for the pool
is empty (i.e., the pool is offline).

Modification:

Parse the URL using the correct path element index.

Result:

The pool name now appears with the OFFLINE state
reported (instead of 'usage', which was the last
element of the API URL).

Target: master
Require-notes: yes
Require-book: no
Acked-by: Olufemi