dcache-view (admin): fix behavior when space data is missing
Motivation:

When data is missing for the pool plots no plot should be
generated.   This does occur for the queue and file lifetime
plots, but not for space data, which is aggregated differently.

Modification:

Add check for definition of the seriesData, not just for
the present of object keys.

Return null when the group data child is missing.

Result:

Space plots behave like the others.

Target: master
Request: 1.4
Require-notes: yes
Require-book: no
Bug: #59
Acked-by: Olufemi