dcache-view (namespace): fix upload request in the context menu
Motivation:

When the upload file is clicked in the context menu,
it result in an Uncaught TypeError because the node
is not attached to the dom when created.

Modification:

Create and dispatch new a custom events called
dv-namespace-upload-files. This event will trigger
a click on the file input element to open a file
picker dialog that allow the user to choose file(s).

Result:

The upload request through context menu now works.

Target: master
Reuest: 1.4
Require-notes: no
Require-book: no
Acked-by: Albert Rossi <arossi@fnal.gov>

Reviewed at https://rb.dcache.org/r/10830/