dcache-view (namespace): fix rendering & lazy-loading issue in view-file
Motivation:

The upgrade of dcache-view to polymer v2.x introduces some
few problems. For example, the lazy loading of data and
rendering of the list of file in a directory is not working
as expected.

Modification:

- Change the scrollTarget of both the iron-list and
iron-scroll-threshold.
- Observe `__counter__` property and set `_isSubsequentRequest`
property to true if `__counter__` is great than 0 else false.
- Restyle view-file element to ensure proper rendering.
- Remove contextmenu eventlistener on the div with id loading
because restyling of view-file ensure that once the loading is
complete, the div display will be set to none.

Result:

The rendering and the lazy loading in view-file is working.

Target: master
Request: 1.4
Require-notes: no
Require-book: no
Acked-by: Albert Rossi <arossi@fnal.gov>

Reviewed at https://rb.dcache.org/r/10847/