dcache-view (namespace): fix the transition from empty to non-empty dir
Motivation:

When a file is dropped or uploaded to an empty directory,
the elevation attribute of the iron-list container is
not updated. Also, this happen when all the content of
the directory is deleted. This issue is as a result of
the new design that is being implemented in dcache-view.

Motivation:

Inside the view-file element, remove all redundancy
codes and use es6 syntax where necessary. Also, add
a new method to handle elevation attribute of
iron-list element container. This is based on whether
the directory is empty or not.

Ensure that this method is called when item(s) is added
or deleted in a directory.

Since event will be dispatch when a new item is added
to the list (in view-file element) and when an item
is deleted; hence, the drop eventlistener callback is
updated to use this feature.

Result:

Fix a bug related to rendering an empty directory when
the last item in directory. Also, this fix the problem
with un-updated elevation attribute whenever an item
is added to an empty directory.

Target: master
Request: 1.4
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/10864/