dcache-view (authentication): ensure backend qos information is set
Motivation:

The backend Qos information suppose to be requested and
set when a user have been authenticated. At the moment
this is not the case. To rectify this problem, the
user-authentication disptaches two events: one for when
the authentication is sucessful and the other for when it
is not. The backend Qos information must be set when the
authentication is sucessful.

Unfortunately, the user-authentication element which is
responsible for user authentication is not attached to any
dom node. Hence, any dispatch event will not be fire unless
they are dispatched to a registered node.

Modification:

1. Disptach all events from the user-authentication
    element with window node as the event target.
2. Rename the events dispatched from user-authentication
    element
2. Adjust all parents of user-authentication element
    accordingly.
3. create a new function called getQosInformation; this
    will set window.CONFIG.qos by creating a new
    QosBackendInformation element if the user is authenticated
    and window.CONFIG.qos is not yet set.
4. Add an eventlistener for sucessful authentication in dv.js
5. Call the new function (getQosInformation) when the page is
    loaded the first time and also when the authentication
    is sucessful.

Result:

The window.CONFIG.qos which quite a few elements used, will
always be set when the user is authenticated.

Target: master
Request: 1.4
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/10865/

(cherry picked from commit 5325ed9242f9f71b816131ac093bbdc17b7ad180)