dcache-view (admin): upgrade dependency to vaadin-grid 4.1.7
Motivation:

Using the 3.0.2 version of the vaadin-grid components, Firefox
performs very badly, particularly in (a) scrolling, which is
very jittery, and (b) inability to load the pool queues page
which uses dom-repeat to build columns dynamically.

While there has been no acknowledgment by the vaadin developers
that this problem can or cannot be solved using version 3.0.2,
it was suggested that we migrate to 4.1.7, which is a version
more in tune with Polymer 2.0.

Modification:

The changes necessary for this upgrade are as follows:

1.  The new grid component implements a themable mixin, and
    and has removed the previous styling attribute mixins.
    Thus all custom style definitions for the vaadin grid
    elements have been removed from the dcache-view elements
    and have been replaced by a single theme component.

2.  The new vaadin-grid no longer implements the cell-click behavior.
    This means that that functionality had to be replaced.
    We have done this, where necessary, by providing a window
    function that sends a CustomEvent, and listeners for those
    events which set state as necessary.

3.  A major change in the table provider behavior seems to
    entail the fact that the provider function is called
    AS EACH FILTER ELEMENT IS ADDED TO THE SHADOW DOM. (WHY?)
    I have asked if there is a way to assign the function
    after the entire DOM is built (it is done in connectedCallback
    at present), and am awaiting response.  If this cannot
    be solved more cleanly, then the current code will
    have to be adopted.  It passes in the number of filters
    to expect in a ready state, and ignores all calls before
    that number is reached.
    Otherwise, there are problems with the table cache because
    multiple async fetches using the same offset but without
    clearing the cache take place.

Because of the new styling, I've also made a few minor adjustments
to the look of the selection and link-group pages.

Result:

The move to 4.1.7 resolves the Firefox issues and brings us
in line with the more modern implementation of vaadin-grid.

Target: master
Request: 1.4
Acked-by: Olufemi