dcache-view (namespace): fix path request in view-file
Motivation:

When the current path of the listed directory is
requested, an undefined respond is returned. This
was due to the fact that the method handling the
request was not properly bind.

Modification:

Properly bind the method handling the request for
current path to the element that owns the property.

Result:

Current path is now send when requested.

Target: master
Request: 1.4
Require-notes: no
Require-book: no
Acked-by: Albert Rossi <arossi@fnal.gov>

Reviewed at https://rb.dcache.org/r/10952/