dcache-view: fix regression regarding the processing of click events on vaadin-grid tables
Motivation:

Patch #10862 commit 2254e597201e99c212c1cd88cb8f8cfd4f8910c1
introduced an upgrade to Vaadin Grid 4.1.7 in order to solve
several issues.

While testing for correct behavior regarding those problems
was done, full regression testing was evidently not conducted.

As a result, a change affecting the handling of click events
on the table was not detected.  Because of this, four views
currently do not response to clicks in order to open file or
pool dialogs.

Modifcation:

Remove the no-longer supported 'on-active-item-changed'
property on the table, define the activeItem property
and its observer explicitly.

Also fixed several small undetected bugs in the related
methods.

Result:

Clicking on the designated columns on the tables now
properly opens the file and/or pool dialogs.

Target: master
Request: 1.4
Acked-by: Olufemi