dcache-view (namespace): style file-icon when list-row is selected
The background and text color of selected file are changed
to indicate to the end-user that the file is selected.
Sadly, this styling doesn't apply to the file-icons' and
this make it less appealing.

Modification:

Add selected attribute to file-icon tag and when set to true
the color of the file-icon element change to white.

Result:

Better styling.

Target: master
Request: 1.4
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan

Reviewed at https://rb.dcache.org/r/10981/